refactor(ExternalLink): make ExternalLink file entrypoint for imports and handle platform specific logic internally
