--- app/components/ExternalLink.web.tsx
@@ -1,1 +1,0 @@
-export { default as ExternalLink } from './ExternalLink.base'
