+++ app/components/onboard/Welcome.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import { Action } from 'redux'
-import ExternalLink from '../ExternalLink.TARGET_PLATFORM'
+import ExternalLink from '../ExternalLink'
 import WelcomeTemplate from './WelcomeTemplate.TARGET_PLATFORM'
 import { DISCORD_URL, GITHUB_ORG_URL } from '../../constants'
