+++ app/components/platformSpecific/ExternalLink.electron.tsx
@@ -1,0 +1,6 @@
+import { shell } from 'electron'
+
+export const onClick = (e: React.MouseEvent, href: string) => {
+  e.preventDefault()
+  shell.openExternal(href)
+}
