+++ app/components/platformSpecific/ExternalLink.web.tsx
@@ -1,0 +1,1 @@
+export const onClick = () => undefined
