+++ app/components/workbench/NotInNamespace.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 
-import ExternalLink from '../ExternalLink.TARGET_PLATFORM'
+import ExternalLink from '../ExternalLink'
 import WorkbenchLayout from './layouts/WorkbenchLayout'
 import DisabledComponentList from './DisabledComponentList'
