+++ app/components/workbench/components/StructureEditor.tsx
@@ -14,7 +14,7 @@ import hasParseError from '../../../utils/hasParseError'
 import SpinnerWithIcon from '../../chrome/SpinnerWithIcon'
 import LabeledStats from '../../item/LabeledStats'
 import Schema from '../../structure/Schema'
-import ExternalLink from '../../ExternalLink.TARGET_PLATFORM'
+import ExternalLink from '../../ExternalLink'
 import FormatConfigEditor from '../../structure/FormatConfigEditor'
 import ParseError from '../ParseError'
 import { connectComponentToProps } from '../../../utils/connectComponentToProps'
